Fix inert analysis menu items and the grid->network String zeroing bug
- N-Cliques, Prestige and Full Analysis menu items were never wired to
  their action listeners (commented out in 2.1.2): clicking them did
  nothing. Listeners restored (cliques, prestige uncommented; full
  analysis added).
- MainFrame.updateNetwork pushed raw String grid cells into the network
  via setObjectValue, which cast them to Float - a ClassCastException per
  cell, silently swallowed, zeroing the entire matrix on every grid-to-
  network update (pre-existing 2.1.2 data-loss bug, visible in the new
  logging). Actor.setObjectEmissionsValue now parses Number and String
  sources; setObjectValue passes values through. Debug prints removed.
- 3 new regression tests (43 total)

+++ src/main/java/com/bentza/sna/net/Network.java
@@ -4,7 +4,7 @@ import com.bentza.sna.AgnaLog;
 import com.bentza.sna.gui.AgnaTableModel;
 import java.util.Vector;
 
-        public class Network
+public class Network
     {
     private Vector all_nodes;
 
@@ -261,7 +261,7 @@ import java.util.Vector;
             }
         try
             {
-            tmp_node.setObjectEmissionsValue((Float) tmp_value, j);
+            tmp_node.setObjectEmissionsValue(tmp_value, j);
             } catch (Exception e) {
       AgnaLog.warn("suppressed exception", e);
       }
